+++ src/presenter/presenter-main.js
@@ -1,11 +1,8 @@
-import { render, replace } from '../framework/render.js';
+import { render } from '../framework/render.js';
 import Sorting from '../view/sorting.js';
 import EventsList from '../view/events-list.js';
-import EventsListItem from '../view/events-list-item.js';
-import Event from '../view/event.js';
-import EditablePoint from '../view/editable-point.js';
 import NoEvents from '../view/no-events.js';
-import { isEscapeKey } from '../utils/common.js';
+import PresenterPoint from './presenter-point.js';
 
 export default class PresenterMain {
   #presenterContainer = null;
@@ -44,51 +41,12 @@ export default class PresenterMain {
   }
 
   #renderPoint(point) {
-    const escKeyDownHandler = (evt) => {
-      if (isEscapeKey(evt)) {
-        evt.preventDefault();
-        replaceFormToEvent();
-        document.removeEventListener('keydown', escKeyDownHandler);
-      }
-    };
-    const destinationId = point.destination;
-    const destination = this.#destinationModel.getById(destinationId);
-    const type = point.type;
-    const offers = this.#offersModel.getByType(type);
-    const list = this.#eventsListComponent.element;
-    const listItem = new EventsListItem();
-    const form = new EditablePoint({
-      point,
-      offers,
-      destination,
-      onClick: () => {
-        replaceFormToEvent();
-        document.removeEventListener('keydown', escKeyDownHandler);
-      },
-      onFormSubmit: () => {
-        replaceFormToEvent();
-        document.removeEventListener('keydown', escKeyDownHandler);
-      },
-    });
-    const event = new Event({
-      point,
-      offers,
-      destination,
-      onClick: () => {
-        replaceEventToForm();
-        document.addEventListener('keydown', escKeyDownHandler);
-      }
+    const presenterPoint = new PresenterPoint({
+      destinationModel: this.#destinationModel,
+      offersModel: this.#offersModel,
+      eventsListComponent: this.#eventsListComponent,
     });
 
-    function replaceFormToEvent() {
-      replace(event, form);
-    }
-
-    function replaceEventToForm() {
-      replace(form, event);
-    }
-
-    render(listItem, list); //рендерим li
-    render(event, listItem.element); //рендерим точку в li
+    presenterPoint.init(point);
   }
 }
